appwrite/backend_scripts/requirements.txt
annotated-types==0.7.0
appwrite==7.0.1
black==24.10.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
idna==3.10
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.3.0
